erase the client if alredy exist and protect kill in case of faild of write

+++ dataCenter/sending.cpp
@@ -6,7 +6,7 @@
 /*   By: rlarabi <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/14 15:58:14 by rlarabi           #+#    #+#             */
-/*   Updated: 2024/02/20 10:43:29 by rlarabi          ###   ########.fr       */
+/*   Updated: 2024/02/20 16:30:30 by rlarabi          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -126,8 +126,10 @@ void dataCenter::sending(int fd){
             {
                 std::cout << RED << "Client disconnected " <<  fd << std::endl;
                 close(fd);
-                kill(this->clientList[fd].getPidCgi(), SIGKILL);
-                waitpid(this->clientList[fd].getPidCgi(), &status, 0);
+                if (this->clientList[fd].getIsCgiExec()){
+                    kill(this->clientList[fd].getPidCgi(), SIGKILL);
+                    waitpid(this->clientList[fd].getPidCgi(), &status, 0);
+                }
                 return;
             }
 
@@ -141,8 +143,10 @@ void dataCenter::sending(int fd){
             std::cout << GREEN << "Response sent GET [ " << res.getStatusCode() << " " << statusCodeMsgs[res.getStatusCode()] << " ] " << fd << " " << this->clientList[fd].getStartLine().path << RESET << "\n";
             write(fd, header.c_str(), header.length());
             close(fd);
-            kill(this->clientList[fd].getPidCgi(), SIGKILL);
-            waitpid(this->clientList[fd].getPidCgi(), &status, 0);
+            if (this->clientList[fd].getIsCgiExec()){
+                kill(this->clientList[fd].getPidCgi(), SIGKILL);
+                waitpid(this->clientList[fd].getPidCgi(), &status, 0);
+            }
             return ;
         }
         if (res.getStatusCode() != 200){
@@ -153,8 +157,10 @@ void dataCenter::sending(int fd){
                     res.setIsHeaderSend(true);
                     write(fd, "HTTP/1.1 403 Forbidden\r\nContent-Type: text/html\r\n\r\nERROR PERMISSION", 67);
                     close(fd);
-                    kill(this->clientList[fd].getPidCgi(), SIGKILL);
-                    waitpid(this->clientList[fd].getPidCgi(), &status, 0);
+                    if (this->clientList[fd].getIsCgiExec()){
+                        kill(this->clientList[fd].getPidCgi(), SIGKILL);
+                        waitpid(this->clientList[fd].getPidCgi(), &status, 0);
+                    }
                     return ;
                 }
             }
@@ -188,8 +194,10 @@ void dataCenter::sending(int fd){
     {
         std::cout << RED << "Client disconnected " <<  fd << std::endl;
         close(fd);
-        kill(this->clientList[fd].getPidCgi(), SIGKILL);
-        waitpid(this->clientList[fd].getPidCgi(), &status, 0);
+        if (this->clientList[fd].getIsCgiExec()){
+            kill(this->clientList[fd].getPidCgi(), SIGKILL);
+            waitpid(this->clientList[fd].getPidCgi(), &status, 0);
+        }
         return;
     }
     if (res.getIsResponseSent()){
@@ -200,8 +208,10 @@ void dataCenter::sending(int fd){
         std::cout << "Response sent " << this->clientList[fd].getStartLine().method << " [ " << res.getStatusCode() << " " << statusCodeMsgs[res.getStatusCode()] << " ] " << fd << " " << this->clientList[fd].getStartLine().path << "\n";
         std::cout << RESET;
         close(fd);
-        kill(this->clientList[fd].getPidCgi(), SIGKILL);
-        waitpid(this->clientList[fd].getPidCgi(), &status, 0);
+        if (this->clientList[fd].getIsCgiExec()){
+            kill(this->clientList[fd].getPidCgi(), SIGKILL);
+            waitpid(this->clientList[fd].getPidCgi(), &status, 0);
+        }
         res.getFilePath().close();
         res.getFilePathError().close();
